+++ test/unit/terrareg/test_version_constraint.py
@@ -45,6 +45,18 @@ class TestVersionConstraint(TerraregUnitTest):
         ('=1.2.3', '0.2.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
         ('=1.2.3', '1.1.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
         ('=1.2.3', '1.2.2', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('=1.2', '1.1.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('=1', '1.2.2', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+
+        ('1.2.3', '1.2.3', terrareg.version_constraint.VersionCompatibilityType.COMPATIBLE),
+        ('1.2.3', '2.2.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2.3', '1.3.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2.3', '1.2.4', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2.3', '0.2.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2.3', '1.1.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2.3', '1.2.2', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1.2', '1.1.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
+        ('1', '1.2.2', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
 
         ('!=1.2.3', '1.2.3', terrareg.version_constraint.VersionCompatibilityType.INCOMPATIBLE),
         ('!=1.2.3', '2.2.3', terrareg.version_constraint.VersionCompatibilityType.IMPLICIT_COMPATIBLE),